+++ src/lib/forms/MultiSelect.svelte
@@ -3,9 +3,6 @@
   import CloseButton from '$lib/utils/CloseButton.svelte';
   import { twMerge } from 'tailwind-merge';
   import type { FormSizeType, SelectOptionType } from '../types';
-  import createEventDispatcher from '$lib/utils/createEventDispatcher';
-
-  const dispatch = createEventDispatcher();
 
   export let items: SelectOptionType<any>[] = [];
   export let value: (string | number)[] = [];
@@ -26,7 +23,7 @@
 
   // Dropdown
   let multiSelectDropdown: string;
-  $: multiSelectDropdown = twMerge('absolute z-50 p-3 flex flex-col gap-1 max-h-64 bg-white border border-gray-300 dark:bg-gray-700 dark:border-gray-600 left-0 top-[calc(100%+1rem)] rounded-lg cursor-pointer overflow-y-scroll w-full', dropdownClass);
+  $: multiSelectDropdown = twMerge('absolute z-50 p-3 flex flex-col gap-1 max-h-64 bg-white border border-gray-300 dark:bg-gray-700 dark:border-gray-600 start-0 top-[calc(100%+1rem)] rounded-lg cursor-pointer overflow-y-scroll w-full', dropdownClass);
 
   // Items
   const itemsClass: string = 'py-2 px-3 rounded-lg text-gray-600 hover:text-gray-600 dark:text-gray-400 hover:bg-gray-100 dark:hover:text-gray-300 dark:hover:bg-gray-600';
@@ -52,6 +49,10 @@
     }
   };
 
+  function create_custom_event(type: string, detail: any, { bubbles = false, cancelable = false } = {}) {
+    return new CustomEvent(type, { detail, bubbles, cancelable });
+  }
+
   function init(node: HTMLSelectElement, value: any) {
     const inital = value; // hack for below
     return {
@@ -59,8 +60,8 @@
         selectItems = items.filter((x) => value.includes(x.value));
         // avoid initial event emitting
         if (value !== inital) {
-          dispatch('change', node, selectItems);
-          dispatch('input', node, selectItems);
+          node.dispatchEvent(create_custom_event('input', selectItems));
+          node.dispatchEvent(create_custom_event('change', selectItems));
         }
       }
     };
@@ -68,7 +69,7 @@
 </script>
 
 <!-- Hidden select for form submission -->
-<select use:init={value} {...$$restProps} {value} hidden multiple>
+<select use:init={value} {...$$restProps} {value} hidden multiple on:change on:input>
   {#each items as { value, name }}
     <option {value}>{name}</option>
   {/each}
@@ -86,12 +87,12 @@
       {/each}
     {/if}
   </span>
-  <div class="flex ml-auto gap-2 items-center">
+  <div class="flex ms-auto gap-2 items-center">
     {#if selectItems.length}
       <CloseButton {size} on:click={clearAll} color="none" class="p-0 focus:ring-gray-400" />
     {/if}
     <div class="w-[1px] bg-gray-300 dark:bg-gray-600" />
-    <svg class="cursor-pointer h-3 w-3 ml-1 text-gray-800 dark:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 10 6">
+    <svg class="cursor-pointer h-3 w-3 ms-1 text-gray-800 dark:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 10 6">
       <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d={show ? 'm1 5 4-4 4 4' : 'm9 1-4 4-4-4'} />
     </svg>
   </div>