+++ src/lib/sidebar/SidebarBrand.svelte
@@ -3,8 +3,8 @@
   import type { SiteType } from '../types';
 
   export let site: SiteType;
-  export let aClass: string = 'flex items-center pl-2.5 mb-5';
-  export let imgClass: string = 'h-6 mr-3 sm:h-7';
+  export let aClass: string = 'flex items-center ps-2.5 mb-5';
+  export let imgClass: string = 'h-6 me-3 sm:h-7';
   export let spanClass: string = 'self-center text-xl font-semibold whitespace-nowrap dark:text-white';
 </script>
 
@@ -18,7 +18,7 @@
 [Go to docs](https://flowbite-svelte.com/)
 ## Props
 @prop export let site: SiteType;
-@prop export let aClass: string = 'flex items-center pl-2.5 mb-5';
-@prop export let imgClass: string = 'h-6 mr-3 sm:h-7';
+@prop export let aClass: string = 'flex items-center ps-2.5 mb-5';
+@prop export let imgClass: string = 'h-6 me-3 sm:h-7';
 @prop export let spanClass: string = 'self-center text-xl font-semibold whitespace-nowrap dark:text-white';
 -->