+++ src/lib/toast/Toast.svelte
@@ -25,10 +25,10 @@
   export let align: boolean = true;
 
   const positions = {
-    'top-left': 'absolute top-5 left-5',
-    'top-right': 'absolute top-5 right-5',
-    'bottom-left': 'absolute bottom-5 left-5',
-    'bottom-right': 'absolute bottom-5 right-5',
+    'top-left': 'absolute top-5 start-5',
+    'top-right': 'absolute top-5 end-5',
+    'bottom-left': 'absolute bottom-5 start-5',
+    'bottom-right': 'absolute bottom-5 end-5',
     none: ''
   };
 
@@ -79,7 +79,6 @@
 @prop export let dismissable: boolean = true;
 @prop export let color: 'primary' | 'gray' | 'red' | 'yellow' | 'green' | 'blue' | 'indigo' | 'purple' | 'orange' | 'none' = 'primary';
 @prop export let position: 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right' | 'none' = 'none';
-@prop export let open: boolean = true;
 @prop export let divClass: string = 'w-full max-w-xs p-4 text-gray-500 bg-white shadow dark:text-gray-400 dark:bg-gray-800 gap-3';
 @prop export let defaultIconClass: string = 'w-8 h-8';
 @prop export let contentClass: string = 'w-full text-sm font-normal';