+++ src/lib/utils/TransitionFrame.svelte
@@ -38,3 +38,13 @@
 {:else}
   <Frame {...$$restProps}><slot {close} /></Frame>
 {/if}
+
+<!--
+@component
+[Go to docs](https://flowbite-svelte.com/)
+## Props
+@prop export let transition: TransitionFunc = fade;
+@prop export let params: object = {};
+@prop export let open: boolean = true;
+@prop export let dismissable: boolean = false;
+-->