+++ src/routes/landing/Hero.svelte
@@ -18,8 +18,9 @@
           <div class="mt-4 md:mt-5 sm:flex justify-center sm:justify-start">
             <div class="flex flex-row mx-0 gap-4 sm:gap-6 items-center max-w-2xl">
               <CopyCliboardInput />
-              <Button size="lg" class="gap-2 max-w-sm md:w-fit whitespace-nowrap hover:bg-primary-800" href="/docs/pages/quickstart" color="primary">
-                Get started <ArrowRight />
+              <Button size="lg" class="space-x-2 rtl:space-x-reverse rtl:space-x-reverse max-w-sm md:w-fit whitespace-nowrap hover:bg-primary-800" href="/docs/pages/quickstart" color="primary">
+                <span>Get started</span>
+                <ArrowRight />
               </Button>
             </div>
           </div>
@@ -28,10 +29,10 @@
 
       <div class="p-0 hidden xl:block">
         <div class="relative dark:hidden">
-          <img class="max-w-xl -mr-7" src="/images/gallery.png" alt="Header" />
+          <img class="max-w-xl -me-7" src="/images/gallery.png" alt="Header" />
         </div>
         <div class="hidden relative dark:block">
-          <img class="max-w-xl -mr-7" src="/images/gallery-dark.png" alt="Header" />
+          <img class="max-w-xl -me-7" src="/images/gallery-dark.png" alt="Header" />
         </div>
       </div>
     </div>