+++ src/routes/utils/AlgoliaSearch.svelte
@@ -19,4 +19,4 @@
   });
 </script>
 
-<div bind:this={searchContainer} id="docsearch" class="ml-3 xl:pl-4" />
+<div bind:this={searchContainer} id="docsearch" class="ms-3 xl:ps-4" />